Rename MemcachedIndexCacheConfig to MemcachedClientConfig and move it to separate file.
